02-streaming/terraform/scripts-hydrated/requirements.txt
streamlit
pandas
numpy
google-cloud-bigquery
db-dtypes